Installs the jquery redactor plugin
All textfields with the redactor-standard class will now be transformed
into a redactor textfield.

This includes jquery migration, since the redactor plugin uses
functions that's been deprecated in jquery 1.9.1.

It also includes the Sanitizer gem, and a String#sanitize extension for
sanitizing the output from the html generated by the redactor wysiwyg.